+++ src/main.rs
@@ -30,13 +30,15 @@ impl Command {
             Command::Build {
                 calls,
                 observations,
+                min_prop_present,
                 output,
             } => {
                 utils::create_output_dir(output)?;
                 let targets = get_targets(calls)?;
                 let mut graphs = HashMap::new();
                 for target in targets {
-                    let variant_graph = VariantGraph::build(calls, observations, &target)?;
+                    let variant_graph =
+                        VariantGraph::build(calls, observations, &target, *min_prop_present)?;
                     if !variant_graph.is_empty() {
                         graphs.insert(target, variant_graph);
                     }
@@ -178,7 +180,7 @@ impl Command {
                     }
                 }
             }
-            Command::Show {
+            Command::Plot {
                 input,
                 format,
                 output,